extract shared delta computation and SMT helpers from store implementations
Move compute_vault_delta and compute_storage_delta from idxdb-store and sqlite-store into rust-client/src/store/delta.rs as shared functions. Add try_get_asset_and_witness helper to AccountSmtForest to eliminate duplicated error handling in both stores.
