+++ crates/rust-client/src/store/mod.rs
@@ -64,6 +64,9 @@ pub(crate) mod data_store;
 mod errors;
 pub use errors::*;
 
+mod delta;
+pub use delta::{compute_storage_delta, compute_vault_delta};
+
 mod smt_forest;
 pub use smt_forest::AccountSmtForest;
